Bug fixes
Fixed a bug where it would crash if a contact has a quote in their name.
Fixed a bug where it would not set a message once it is time to send.
Fixed a bug where it would send multiple messages instead of one.
Application now resets settings whenever time or timezone is manually changed.